+++ src/handle/back_dialog.js
@@ -19,14 +19,12 @@ module.exports = () => {
     var currentModel = getCurrentDialogModel();
     var prevModel = getPrevDialogModel();
     var currentPageModel = getCurrentPageModel();
-    var hasPrevDialog = !1;
 
     // for result
     if (currentModel.forResult) {
         // has prev dialog model
         if (prevModel) {
             prevModel.dialog.onResult && prevModel.dialog.onResult(currentModel.dialog.__orchids__result || {});
-            hasPrevDialog = !0;
         }
         else if (currentPageModel) {
             currentPageModel.page.onResult && currentPageModel.page.onResult(currentModel.dialog.__orchids__result || {});
@@ -35,7 +33,7 @@ module.exports = () => {
 
     // destroy or hide
     currentModel.singleton ? currentModel.dialog.__orchids__hide(!0) : currentModel.dialog.__orchids__destroy();
-    hasPrevDialog ? prevModel.dialog.__orchids__show() : currentPageModel && currentPageModel.page.__orchids__show();
+    prevModel ? prevModel.dialog.__orchids__show() : currentPageModel && currentPageModel.page.__orchids__show();
 
     deleteCurrentDialogModel();
 
@@ -49,11 +47,27 @@ module.exports = () => {
             setTimeout(() => {
                 currentModel.dialog.el.remove();
                 currentModel.dialog.afterDestroy();
+                prevModel ? prevModel.dialog.afterShow() : currentPageModel && currentPageModel.page.afterShow();
             }, vars.animateTime);
         else {
             // no animation
             currentModel.dialog.el.remove();
             currentModel.dialog.afterDestroy();
+            prevModel ? prevModel.dialog.afterShow() : currentPageModel && currentPageModel.page.afterShow();
+        }
+    }
+    else {
+        if (currentModel.dialog.option.animate) {
+            // has animation
+            setTimeout(() => {
+                currentModel.dialog.afterHide();
+                prevModel ? prevModel.dialog.afterShow() : currentPageModel && currentPageModel.page.afterShow();
+            }, vars.animateTime)
+        }
+        else {
+            // no animation
+            currentModel.dialog.afterHide();
+            prevModel ? prevModel.dialog.afterShow() : currentPageModel && currentPageModel.page.afterShow();
         }
     }
 };