+++ src/handle/start_dialog_inner.js
@@ -49,10 +49,25 @@ module.exports = (name, data, forResult, prepareResultData) => {
 
         // has initialized before
         if (singletonModel) {
-            container.dialogModels[vars.idPrefix + singletonModel.id] = makeDialogModel(name, forResult, singletonModel.dialog, !0);
-
             forResult ? singletonModel.dialog.__orchids__show(!0, !0, prepareResultData) : singletonModel.dialog.__orchids__show(!0);
 
+            container.dialogModels[vars.idPrefix + singletonModel.id] = makeDialogModel(name, forResult, singletonModel.dialog, !0);
+
+            // call dialog's afterShow
+            if (!singletonModel.dialog.option.animate) {
+                if (currentDialogModel) currentDialogModel.dialog.afterHide();
+                else if (currentPageModel) currentPageModel.page.afterHide();
+
+                singletonModel.dialog.afterShow();
+            }
+            else {
+                setTimeout(() => {
+                    if (currentDialogModel) currentDialogModel.dialog.afterHide();
+                    else if (currentPageModel) currentPageModel.page.afterHide();
+
+                    singletonModel.dialog.afterShow();
+                }, vars.animateTime);
+            }
             return;
         }
     }
@@ -75,13 +90,21 @@ module.exports = (name, data, forResult, prepareResultData) => {
     container.dialogModels[vars.idPrefix + dialogOption.dialogId] = makeDialogModel(name, forResult, instance, dialogOption.singleton);
 
     // call dialog's afterCreate
-    if (!instance.option.animate) instance.afterCreate();
+    if (!instance.option.animate) {
+        if (currentDialogModel) currentDialogModel.dialog.afterHide();
+        else if (currentPageModel) currentPageModel.page.afterHide();
+
+        instance.afterCreate();
+    }
     else {
         // show dialog, delay 100 ms to guarantee the animation  is ok, and 0 is not ok
         setTimeout(() => {
             instance.el.classList.add('orchids-active');
 
             setTimeout(() => {
+                if (currentDialogModel) currentDialogModel.dialog.afterHide();
+                else if (currentPageModel) currentPageModel.page.afterHide();
+
                 instance.afterCreate();
             }, vars.animateTime);
         }, vars.animateDelayTime);